requirements.txt
pandas~=2.2.3
scikit-learn~=1.6.1
nltk==3.8.1
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2
flask==3.0.0
flask-cors==4.0.0
transformers~=4.52.3
torch~=2.7.0
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
lxml_html_clean==0.1.1
python-dotenv==1.0.0
numpy~=1.26.4
datasets~=3.6.0
plotly~=6.1.2